+++ sdk/spring/spring-cloud-azure-service/src/main/java/com/azure/spring/cloud/service/implementation/keyvault/certificates/CertificateClientBuilderFactory.java
@@ -88,6 +88,12 @@ public class CertificateClientBuilderFactory extends AbstractAzureHttpClientBuil
         PropertyMapper map = new PropertyMapper();
         map.from(certificateClientProperties.getEndpoint()).to(builder::vaultUrl);
         map.from(certificateClientProperties.getServiceVersion()).to(builder::serviceVersion);
+        map.from(certificateClientProperties.isDisableChallengeResourceVerification()).to(
+            disabled -> {
+                if (disabled) {
+                    builder.disableChallengeResourceVerification();
+                }
+            });
     }
 
     @Override